vcl: migrate hardware font activation and metric retrieval to FontController
Continues the decoupling of OutputDevice from font management by delegating
hardware-specific synchronization and collection-based metric retrieval to
the FontController.

Changes:
- Added FontController::ActivateFontOnDevice to encapsulate the
  direct SalGraphics::SetFont call.
- Added FontController::GetFontMetricFromCollection to handle the lazy
  initialization and flat-view retrieval of font metrics from the
  PhysicalFontCollection.
- Updated OutputDevice::InitFont and GetFontMetricFromCollection to
  delegate to the controller, ensuring the controller is lazily
  initialized if needed.
- Fixed a potential null pointer dereference in OutputDevice by
  adding safety checks for mpFontController and mpGraphics.
- Updated PhysicalFontCollection and related unit tests to use
  size_t for Count() to ensure type consistency and avoid
  unnecessary casting.

This moves OutputDevice further toward being a passive consumer of
font services rather than a manager of font engine state.

+++ vcl/qa/cppunit/physicalfontcollection.cxx
@@ -94,20 +94,20 @@ public:
 void VclPhysicalFontCollectionTest::testShouldCreateAndAddFontFamilyToCollection()
 {
     vcl::font::PhysicalFontCollection aFontCollection;
-    CPPUNIT_ASSERT_EQUAL_MESSAGE("Empty font collection", static_cast<int>(0),
+    CPPUNIT_ASSERT_EQUAL_MESSAGE("Empty font collection", static_cast<size_t>(0),
                                  aFontCollection.Count());
 
     // please note that fonts created this way are NOT normalized and will not be found if you search for them in the collection
     vcl::font::PhysicalFontFamily* pFontFamily
         = aFontCollection.FindOrCreateFontFamily(u"Test Font Family Name"_ustr);
     CPPUNIT_ASSERT_EQUAL_MESSAGE("Does not have only one font family in collection",
-                                 static_cast<int>(1), aFontCollection.Count());
+                                 static_cast<size_t>(1), aFontCollection.Count());
     CPPUNIT_ASSERT_EQUAL_MESSAGE("Font family name not correct", u"Test Font Family Name"_ustr,
                                  pFontFamily->GetSearchName());
 
     vcl::font::PhysicalFontFamily* pFontFamily2
         = aFontCollection.FindOrCreateFontFamily(u"Test Font Family Name"_ustr);
-    CPPUNIT_ASSERT_EQUAL_MESSAGE("Still only one font family in collection", static_cast<int>(1),
+    CPPUNIT_ASSERT_EQUAL_MESSAGE("Still only one font family in collection", static_cast<size_t>(1),
                                  aFontCollection.Count());
     CPPUNIT_ASSERT_EQUAL_MESSAGE("Font family name not correct", u"Test Font Family Name"_ustr,
                                  pFontFamily2->GetSearchName());
